Remove warnings of type unneeded conversions. Fix polynomial evaluation with fixed_terms and hypercube values in the same vec.

+++ src/polynomial/monomial.rs
@@ -50,7 +50,7 @@ impl<const P: u64> Monomial<P> {
 
         for (index, exponent) in self.exponents.iter().enumerate() {
             if fixed_value == index as u64 {
-                if *exponent == 1 as usize {
+                if *exponent == 1 {
                     result_exponent = 1;
                 }
             } else {
